requirements.txt
Flask
Flask-Cors
google-generativeai
pyttsx3
